# Speedup autoloading with APCu class loader