--- a/tests/Composer/Test/Autoload/Fixtures/autoload_real_files_by_dependency.php
+++ b/tests/Composer/Test/Autoload/Fixtures/autoload_real_files_by_dependency.php
@@ -10,6 +10,8 @@ class ComposerAutoloaderInitFilesAutoloadOrder
     {
         if ('Composer\Autoload\ClassLoader' === $class) {
             require __DIR__ . '/ClassLoader.php';
+        } elseif ('Composer\Autoload\ApcuClassLoader' === $class) {
+            require __DIR__ . '/ApcuClassLoader.php';
         }
     }