--- a/tests/Composer/Test/Autoload/Fixtures/autoload_real_include_path.php
+++ b/tests/Composer/Test/Autoload/Fixtures/autoload_real_include_path.php
@@ -10,6 +10,8 @@ class ComposerAutoloaderInitIncludePath
     {
         if ('Composer\Autoload\ClassLoader' === $class) {
             require __DIR__ . '/ClassLoader.php';
+        } elseif ('Composer\Autoload\ApcuClassLoader' === $class) {
+            require __DIR__ . '/ApcuClassLoader.php';
         }
     }