The fastest autoloader we can currently get with composer is when dumping the class map in authoritative mode: this reduces autoloading to an `isset()` check, one of the fastest in PHP.
But this autoloader has one drawback: it doesn't scale with the number of classes. Even with opcache enabled, the generated class map being a php array, it becomes slower when you add more classes to it. It also consumes more memory.

That made me wonder if using APCu could help build an even better autoloader.

This PR adds an ApcuClassLoader that wraps the regular ClassLoader by inheritance so that only APCu is used when the cache is warm, and the regular autoloader is lazily initialized when needed to warm up the APCu cache.

I did not yet wire any on/off way to configure which loader to use. Just change yourself the `$apcuLoader` property if you want to try the patch for now.

Still, I wanted to share with you the following Blackfire profile:
https://blackfire.io/profiles/compare/ace900a5-4936-4998-871d-680bbf744177/graph

This profile has been generated by taking the blackfire classmap (~8300 lines) and loading ~430 classes from it, this number being quite typical for rendering a single page/response in a Symfony app.

The results are quite interesting:
- loading the classmap goes from 2.66 ms and 2.56 MB to zero
- overall loading time is reduced by 1.59 ms and 2.54 MB!

On the memory side, we have a clear win.
On the time side, 2.66 ms removal from not-loading the classmap are turned into a lower 1.59 ms general perf increase. Logically, this could mean that *once the classmap is loaded*, using it is faster that calling apcu_fetch, but since we have to load this classmap anyway, the win is still a clear win in the typical web request use case where only a short excerpt of all existing classes are loaded at once.

The only remaining question to me now is: how should we plug this into composer configuration system.
Some ideas:
- add a command line option?
- `ext-apcu` must-be in the dependencies?
- both? Other way?