@beberlei 2016-04-10
@nicolas-grekas Can you make the same comparison on PHP 7? I would be interested how much of the measured benefit will go away, since hashtables were optimized so much.