@omansour 2016-04-10
brilliant !!! 

:dancers: 

I think detecting ext-apcu and using the apcu loader automatically has my preference. But maybe it add some small overhead too. So ext-apcu will be suggested