@jderusse 2016-04-10
Given the cache is shared, what happens when the same machine run 2 applications using different versions of the same class ?