@jderusse 2016-04-10
Oh. I miss the prefix thing. 

Is the prefix regenerated after each install? For instance after that a dependency is upgrade/downgraded...