@stof 2016-04-10
@jderusse currently, it is regenerated randomly for each autoload dump. See my comment about reusing the existing configurable setting (which still defaults to randomly generating a value)