@nicolas-grekas 2016-04-10
In reality, the suffix is not always regenerated, so that reusing it wouldn't always invalidate the cache when required, that's why I did it this way.