@JanTvrdik 2016-04-11
> ApcuClassLoader that wraps the regular ClassLoader by inheritance

Why not use composition instead of inheritance? Why bother loading `ClassLoader` at all when APCu cache is ready?