@nicolas-grekas 2016-04-11
@JanTvrdik @omansour because `$loader instanceof ClassLoader` must stay true, and also because we need lazy-initialization.