@nicolas-grekas 2016-04-11
Closing in favor of #5174